requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Video processing
opencv-python-headless==4.8.1.78
ffmpeg-python==0.2.0
pillow==10.1.0
numpy==1.24.3

# AI/ML
ollama==0.1.7

# Testing
pytest==7.4.3
pytest-bdd==6.1.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1

# Device control
requests==2.31.0
aiohttp==3.9.1

# State management
redis==5.0.1
python-dotenv==1.0.0

# Utilities
pyyaml==6.0.1
loguru==0.7.2
jinja2==3.1.2
tenacity==8.2.3

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0